fix(web): claim handle in a modal at publish need
Drop the inline Studio claim panel. Waiting games show a compact
“Claim handle to publish” control that opens a modal. Header name
links to /creators/:handle once claimed. Still no forced signup slug.
