+++ apps/web/src/i18n/locales/en.json
@@ -853,7 +853,7 @@
     "publishGateTitle": "Claim a handle to publish",
     "editorNeeded": "Building and playtesting do not need a handle — publishing does.",
     "editorReady": "This is how players see you on catalog cards and the player.",
-    "publishNudge": "This game is waiting on review — claim a handle so it can go live.",
+    "publishNudge": "This game is ready to go live. Pick a unique handle for your public page.",
     "editProfile": "Edit profile",
     "done": "Done",
     "previewAria": "Public profile preview",
