+++ apps/web/src/i18n/locales/pl.json
@@ -857,7 +857,7 @@
     "publishGateTitle": "Zajmij handle, żeby opublikować",
     "editorNeeded": "Do budowania i testów handle nie jest potrzebny — do publikacji tak.",
     "editorReady": "Tak widzą Cię gracze na kartach katalogu i w playerze.",
-    "publishNudge": "Ta gra czeka na recenzję — zajmij handle, żeby mogła trafić na stronę.",
+    "publishNudge": "Ta gra jest gotowa do publikacji. Wybierz unikalny handle na swoją publiczną stronę.",
     "editProfile": "Edytuj profil",
     "done": "Gotowe",
     "previewAria": "Podgląd publicznego profilu",
